FURKANB/requirements.txt
Flask
gTTS
google-generativeai
python-dotenv
SpeechRecognition
